pyproject.toml
[tool.poetry]
name = "vjn-discord-order"
version = "0.1.0"
description = ""
authors = ["Dov Devers <renaud-dov.devers@epita.fr>", "Clovis Lefebvre <clovis.lefebvre@epita.fr>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
"discord.py" = "^2.3.2"
PyYAML = "^6.0"
jsonschema = "^4.17.3"
SQLAlchemy = "^2.0.19"
pydantic = "^2.3.0"
psycopg2 = "^2.9.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
